Fix simulator
