--- contracts/exec-caller-by-code-hash/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "exec-caller-by-code-hash"
-version = "0.1.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-ckb-std = { path = "../../", features = ["dummy-atomic"] }
